refactor(pty): extract TerminalInputController for write/submit/stage
- new TerminalInputController owns tryWrite/write/submit/stage/performSubmit and the focus-input side effects
- TerminalProcess keeps the public tryWrite/write/submit/stage methods and the private performSubmit as thin delegating calls
