+++ electron/services/pty/TerminalProcess.ts
@@ -27,6 +27,7 @@ import {
 } from "./types.js";
 import { WriteQueue } from "./WriteQueue.js";
 import { AgentOutputForwarder } from "./AgentOutputForwarder.js";
+import { TerminalInputController } from "./TerminalInputController.js";
 import { events } from "../events.js";
 import { AgentSpawnedSchema } from "../../schemas/agent.js";
 import { destroyPty, type PooledPtyDataHandoff, type PtyPool } from "../PtyPool.js";
@@ -36,23 +37,6 @@ import { handleOscColorQueries } from "./OscResponder.js";
 import { Osc94Parser } from "./Osc94Parser.js";
 import { SynchronizedFrameDetector } from "./SynchronizedFrameDetector.js";
 
-// Extracted modules
-import {
-  normalizeSubmitText,
-  splitTrailingNewlines,
-  supportsBracketedPaste,
-  getSoftNewlineSequence,
-  getSubmitEnterDelay,
-  isBracketedPaste,
-  isFocusReport,
-  delay,
-  BRACKETED_PASTE_START,
-  BRACKETED_PASTE_END,
-  PASTE_THRESHOLD_CHARS,
-  OUTPUT_SETTLE_DEBOUNCE_MS,
-  OUTPUT_SETTLE_MAX_WAIT_MS,
-  OUTPUT_SETTLE_POLL_INTERVAL_MS,
-} from "./terminalInput.js";
 import type { IMarker } from "@xterm/headless";
 import {
   TERMINAL_SESSION_PERSISTENCE_ENABLED,
@@ -68,13 +52,8 @@ import {
 import { TerminalForensicsBuffer } from "./TerminalForensicsBuffer.js";
 import { SemanticBufferManager } from "./SemanticBufferManager.js";
 import { ProcessTreeKiller } from "./ProcessTreeKiller.js";
-import {
-  IdentityWatcher,
-  normalizeShellCommandText,
-  type IdentityWatcherDelegate,
-} from "./IdentityWatcher.js";
+import { IdentityWatcher, type IdentityWatcherDelegate } from "./IdentityWatcher.js";
 import type { SpawnContext } from "./terminalSpawn.js";
-import { logIdentityDebug } from "./identityDebug.js";
 import { computeDefaultTitle, getLiveAgentId } from "./terminalTitle.js";
 import {
   serializeTerminal,
@@ -175,6 +154,7 @@ export class TerminalProcess {
   private identityWatcher!: IdentityWatcher;
 
   private writeQueue!: WriteQueue;
+  private inputController!: TerminalInputController;
   private readonly processTreeKiller: ProcessTreeKiller;
   private readonly lifecycle = new TerminalProcessLifecycle();
 
@@ -451,6 +431,24 @@ export class TerminalProcess {
     });
     this.sessionSnapshotter = this.createSessionSnapshotter();
     this.identityWatcher = new IdentityWatcher(this.createIdentityWatcherDelegate());
+    this.inputController = new TerminalInputController({
+      get id() {
+        return self.id;
+      },
+      get terminalInfo() {
+        return self.terminalInfo;
+      },
+      get analysis() {
+        return self.analysis;
+      },
+      get identityWatcher() {
+        return self.identityWatcher;
+      },
+      get writeQueue() {
+        return self.writeQueue;
+      },
+      logWriteError: (error, context) => self.logWriteError(error, context),
+    });
     this.exitObservers = new TerminalExitObservers({
       id: this.id,
       terminalInfo: this.terminalInfo,
@@ -1093,151 +1091,15 @@ export class TerminalProcess {
    * but broadcast keystrokes are always single chunks so this is fine.
    */
   tryWrite(data: string, traceId?: string): { ok: boolean; error?: NodeJS.ErrnoException } {
-    const terminal = this.terminalInfo;
-    if (terminal.isExited) {
-      return {
-        ok: false,
-        error: Object.assign(new Error("terminal exited"), { code: "EBADF" }),
-      };
-    }
-    if (!terminal.ptyProcess) {
-      return {
-        ok: false,
-        error: Object.assign(new Error("terminal has no pty process"), { code: "EBADF" }),
-      };
-    }
-
-    if (data.length > 512) {
-      // Long payloads queue through chunkInput in write(); we lose precise
-      // per-call failure visibility but that path isn't used by broadcast.
-      this.write(data, traceId);
-      return { ok: true };
-    }
-
-    terminal.lastInputTime = Date.now();
-    if (traceId !== undefined) {
-      terminal.traceId = traceId || undefined;
-    }
-    if (this.analysis.hasMonitor()) {
-      if (isFocusReport(data)) {
-        this.handleFocusInput();
-      } else {
-        this.analysis.notifyInput(data);
-      }
-    }
-
-    try {
-      terminal.ptyProcess.write(data);
-      return { ok: true };
-    } catch (error) {
-      this.logWriteError(error, { operation: "tryWrite", traceId });
-      return { ok: false, error: error as NodeJS.ErrnoException };
-    }
+    return this.inputController.tryWrite(data, traceId);
   }
 
   write(data: string, traceId?: string): void {
-    const terminal = this.terminalInfo;
-    terminal.lastInputTime = Date.now();
-
-    if (terminal.isExited) {
-      return;
-    }
-
-    if (!terminal.ptyProcess) {
-      return;
-    }
-
-    if (traceId !== undefined) {
-      terminal.traceId = traceId || undefined;
-    }
-
-    if (this.analysis.hasMonitor()) {
-      if (isFocusReport(data)) {
-        this.handleFocusInput();
-      } else {
-        this.analysis.notifyInput(data);
-      }
-    }
-
-    const bracketedPaste = isBracketedPaste(data);
-    const seededCommandText = this.identityWatcher.seededCommandText;
-    const isSeededLaunchCommandSubmit =
-      !bracketedPaste &&
-      seededCommandText !== undefined &&
-      /[\r\n]/.test(data) &&
-      normalizeShellCommandText(data) === seededCommandText;
-    // Shell input capture is only meaningless when a live AGENT owns the PTY
-    // (agents have their own input semantics). A plain process badge (npm,
-    // pnpm, docker, etc.) does not change the shell semantics — the shell
-    // is still the direct recipient of typed commands, and the next command
-    // must still be visible to the fallback detector so a follow-up
-    // `pnpm build` can re-identify the badge. #5813
-    const canCaptureShellInput =
-      !bracketedPaste &&
-      (this.terminalInfo.detectedAgentId === undefined || isSeededLaunchCommandSubmit);
-    const submittedCommandText = canCaptureShellInput
-      ? this.identityWatcher.captureInput(data)
-      : undefined;
-    const pendingFallbackIdentity = this.identityWatcher.pendingFallbackIdentity;
-    const isAgentUiPromptResponse =
-      !bracketedPaste &&
-      submittedCommandText === undefined &&
-      pendingFallbackIdentity?.agentType !== undefined &&
-      (!this.identityWatcher.isFallbackCommitted ||
-        this.identityWatcher.hasAgentUiPromptFalsePositive());
-
-    if (!bracketedPaste && /[\r\n]/.test(data)) {
-      if (this.identityWatcher.consumeSuppressSignal()) {
-        // Suppression consumed — performSubmit() armed it for its body+enter sequence.
-      } else if (isAgentUiPromptResponse) {
-        logIdentityDebug(
-          `[IdentityDebug] shell-submit-skip term=${this.id.slice(-8)} reason=agent-ui-prompt`
-        );
-      } else {
-        this.identityWatcher.onShellSubmit(submittedCommandText, {
-          allowWhenAgentDetected: isSeededLaunchCommandSubmit,
-        });
-      }
-      if (isSeededLaunchCommandSubmit) {
-        this.identityWatcher.clearSeededCommandText();
-      }
-    }
-
-    if (bracketedPaste) {
-      try {
-        terminal.ptyProcess.write(data);
-      } catch (error) {
-        this.logWriteError(error, { operation: "write(bracketed-paste)", traceId });
-      }
-      return;
-    }
-
-    if (data.length <= 512) {
-      try {
-        terminal.ptyProcess.write(data);
-      } catch (error) {
-        this.logWriteError(error, { operation: "write(fast-path)", traceId });
-      }
-      return;
-    }
-
-    this.writeQueue.enqueueChunked(data);
+    this.inputController.write(data, traceId);
   }
 
   submit(text: string): void {
-    if (this.terminalInfo.isExited) {
-      return;
-    }
-
-    // Immediately notify activity monitor of the submission so the working
-    // state transitions before the async write sequence in performSubmit().
-    // Without this, the split between body write and Enter write causes the
-    // character-by-character detection in onInput() to miss the submission.
-    if (this.analysis.hasMonitor() && text.trim().length > 0) {
-      this.analysis.notifySubmission();
-    }
-
-    this.writeQueue.submit(text);
+    this.inputController.submit(text);
   }
 
   /**
@@ -1251,91 +1113,11 @@ export class TerminalProcess {
    * by `host.sendToActiveAgent(text, { submit: false })` (#10558).
    */
   stage(text: string): void {
-    const terminal = this.terminalInfo;
-    if (terminal.isExited || !terminal.ptyProcess) {
-      return;
-    }
-    const normalized = normalizeSubmitText(text);
-    const { body } = splitTrailingNewlines(normalized);
-    if (body.length === 0) {
-      return;
-    }
-    terminal.lastInputTime = Date.now();
-    const useBracketedPaste = body.includes("\n") || body.length > PASTE_THRESHOLD_CHARS;
-    if (useBracketedPaste && supportsBracketedPaste(terminal)) {
-      const pasteBody = body.replace(/\n/g, "\r");
-      this.write(`${BRACKETED_PASTE_START}${pasteBody}${BRACKETED_PASTE_END}`);
-    } else if (body.includes("\n")) {
-      this.write(body.replace(/\n/g, getSoftNewlineSequence(terminal)));
-    } else {
-      this.write(body);
-    }
+    this.inputController.stage(text);
   }
 
   private async performSubmit(text: string): Promise<void> {
-    const terminal = this.terminalInfo;
-    terminal.lastInputTime = Date.now();
-
-    if (terminal.isExited) {
-      return;
-    }
-
-    if (!terminal.ptyProcess) {
-      return;
-    }
-
-    // Notify activity monitor at execution time (not just enqueue time) to ensure
-    // the working state transition happens even for queued submissions that execute
-    // after a potential idle transition. Issue #2185.
-    if (this.analysis.hasMonitor() && text.trim().length > 0) {
-      this.analysis.notifySubmission();
-    }
-
-    const normalized = normalizeSubmitText(text);
-    const { body, enterCount } = splitTrailingNewlines(normalized);
-    const enterSuffix = "\r".repeat(enterCount);
-
-    if (body.length === 0) {
-      this.identityWatcher.armSuppressSignal();
-      this.write(enterSuffix);
-      return;
-    }
-
-    const useBracketedPaste = body.includes("\n") || body.length > PASTE_THRESHOLD_CHARS;
-    const useOutputSettle = !supportsBracketedPaste(terminal);
-
-    if (useBracketedPaste && supportsBracketedPaste(terminal)) {
-      const pasteBody = body.replace(/\n/g, "\r");
-      const payload = `${BRACKETED_PASTE_START}${pasteBody}${BRACKETED_PASTE_END}`;
-      this.write(payload);
-    } else {
-      if (body.includes("\n") && !supportsBracketedPaste(terminal)) {
-        const softNewline = getSoftNewlineSequence(terminal);
-        this.write(body.replace(/\n/g, softNewline));
-      } else {
-        this.write(body);
-      }
-    }
-
-    await this.writeQueue.waitForInputWriteDrain();
-
-    if (useOutputSettle) {
-      await this.writeQueue.waitForOutputSettle({
-        debounceMs: OUTPUT_SETTLE_DEBOUNCE_MS,
-        maxWaitMs: OUTPUT_SETTLE_MAX_WAIT_MS,
-        pollMs: OUTPUT_SETTLE_POLL_INTERVAL_MS,
-      });
-    } else {
-      await delay(getSubmitEnterDelay(terminal));
-    }
-
-    if (!this.terminalInfo.ptyProcess) {
-      return;
-    }
-
-    this.identityWatcher.armSuppressSignal();
-    this.identityWatcher.onShellSubmit(body);
-    this.write(enterSuffix);
+    await this.inputController.performSubmit(text);
   }
 
   resize(cols: number, rows: number): void {
@@ -1795,15 +1577,6 @@ export class TerminalProcess {
       // strands the FSM in working (#9875).
       this.activityMonitor?.notifyExternalPromotion();
     }
-  }
-
-  // Side-effects shared by both PTY write paths when xterm forwards a CSI I/O
-  // focus report. Mirrors the resize handler's pattern (notifyResize +
-  // agentOutputTemperature.noteResize): open the ActivityMonitor suppression
-  // window AND invalidate the agentOutputTemperature baseline so the redraw
-  // that follows the focus event is treated as a fresh comparison point.
-  private handleFocusInput(): void {
-    this.analysis.notifyFocus();
   }
 
   /**
